+++ src/components/Button.svelte
@@ -1,12 +1,11 @@
 <script lang=ts>
-  export let primary:boolean
+  export let primary:boolean = false
 </script>
 
-<button class="py-2 px-4 hover:font-bold" on:click class:primary><slot /></button>
+<button class="py-2 px-4 hover:font-italic" on:click class:primary><slot /></button>
 
 <style>
   .primary {
-    box-sizing: content-box;
     background: var(--accent-color);
     border-color: var(--accent-color);
     color: white;