+++ CHANGELOG.md
@@ -9,3 +9,4 @@
   - import 'uno.css' in __layout.svelte
 - add @fontsource/krub
 - change main layout to [flex](https://dev.to/domysee/keeping-the-footer-at-the-bottom-with-css-flexbox-5h5f)
+- button.svelte component